refactor(auction): AuctionApiClient 경로 변경
